Mudanca dos nome de chefao para monstro e retirado o tio de colisoes

+++ TecProgTrabalho/Alien.h
@@ -7,9 +7,9 @@ namespace Entidades
 {
 	class Alien : public Inimigo {
 	private:
-		float tamanho;
+		int tamanho;
 	public:
-		Alien(sf::Vector2f p = sf::Vector2f(1400.f, 500.f), int n = 0, float tam = 1.f);
+		Alien(sf::Vector2f p = sf::Vector2f(1400.f, 500.f), int n = 0, int tam = 1);
 		Alien(const nlohmann::json& dados);
 		~Alien();
 		void salvar();
